pyproject.toml
[project]
name = "ai-employee"
version = "0.1.0"
description = "Personal AI Employee - Your Digital FTE (Full-Time Equivalent)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.193.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.3.0",
    "playwright>=1.58.0",
    "watchdog>=6.0.0",
]
